pyproject.toml
[tool.poetry]
name = "event-stream-processor"
version = "0.1.0"
description = "Register functions to be passed events from a queue"
authors = ["etiology"]

[tool.poetry.dependencies]
python = "^3.9"
confluent-kafka = "^1.7.0"
rich = "^10.9.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
black = "^20.8b1"
mkdocs = "^1.1.2"
mkdocs-material = "^7.1.4"
mkdocstrings = "^0.15.1"
bandit = "^1.7.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
